chore(EMS-3501): get populated application - split up export contract
